# docs/EmpAppAPIDoc.json
{"openapi":"3.1.0","info":{"title":"Employee Management API","description":"API for managing employees","version":"1.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Employee API","description":"Operations related to employees"}],"paths":{"/api/employees/{id}":{"get":{"tags":["Employee API"],"operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeEntity"}}}}}},"put":{"tags":["Employee API"],"operationId":"updateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeEntity"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Employee API"],"operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/employees":{"get":{"tags":["Employee API"],"operationId":"getAllEmployees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeEntity"}}}}}}},"post":{"tags":["Employee API"],"operationId":"addEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeEntity"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/test":{"get":{"tags":["test-swagger-controller"],"summary":"Test endpoint","operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"EmployeeEntity":{"type":"object","properties":{"id":{"type":"string"},"employeeEmailId":{"type":"string"},"employeeFirstName":{"type":"string"},"employeeLastName":{"type":"string"}}}}}}